Allow for direct unsubscribe

+++ list/list-funcs.php
@@ -1,16 +1,18 @@
 <?php
 function subscribe($email) {
+if (! is_writable("subs/$id")) { return false; }
 $id=trim(`head -c 4 /dev/urandom | xxd -p`);
-file_put_contents("subs/$id", time() . " " . $_SERVER["REMOTE_ADDR"] . " " . $email . "\n", FILE_APPEND);
+file_put_contents("subs/$id", gethostbyaddr($_SERVER["REMOTE_ADDR"]) . " " . $email . "\n", FILE_APPEND);
 return $id;
 }
 
 function unsubscribe($id) {
+if (! is_writable("subs/$id")) { return false; }
 $fp = fopen("subs/$id", "r");
-while ($row = fgetcsv($fp, 0, " ")) {
-	$email = $row[2];
-	file_put_contents("usubs.csv", "$row[0] $row[1] $row[2] " . filter_var(trim($_POST[why]), FILTER_SANITIZE_STRING) . "\n", FILE_APPEND);
-}
+$row = fgetcsv($fp, 0, " ");
+$email = end($row);
+$unsubscription_details = implode(' ', array(time(), time() - filemtime("subs/$id"), implode(' ', $row), filter_var(trim($_REQUEST[why]), FILTER_SANITIZE_STRING)));
+file_put_contents("usubs.private", $unsubscription_details . "\n", FILE_APPEND);
 fclose($fp);
 unlink("subs/$id");
 return $email;
